packages/cartrita-v2/py/requirements.txt
# Cartrita V2 - Python Dependencies
# OpenAI and AI libraries
openai>=1.54.0
langchain-openai>=0.1.0
langchain-core>=0.2.0

# FastAPI and web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP and networking
httpx>=0.26.0
requests>=2.31.0
websockets>=12.0

# Computer use and automation
pyautogui>=0.9.54
pillow>=10.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
playwright>=1.44.0
selenium>=4.15.0

# System monitoring
psutil>=5.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.19.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0